playing with es6 classes wip
