+++ client/handlebars/taskTpl.hbs
@@ -2,4 +2,4 @@
 <span class="task-info">Name: <span class="task-name">{{name}}</span></span>
 <span class="task-info">Author: <span class="task-author">{{author}}</span></span>
 <span class="task-info">Description: <span class="task-desc">{{description}}</span></span>
-<button class="edit-task" onclick="showEditDialog(this);">Edit Task</button>
+<button class="edit-task" onclick="showEditDialog(this.tasks);">Edit Task</button>
