+++ client/js/main.js
@@ -1,201 +1,9 @@
 'use strict';
-import * as Api from './api';
+import App from './app';
 
-let tasks, taskTpl,
-    isEditing = false,
-    taskId;
+let app = new App();
 
-getTemplate().then( (data) => { //get hbs template
-  taskTpl = Handlebars.compile(data);
-  getTasks().then( (tasks) => { //get tasks from API
-    for (let i = 0 ; i < tasks.length ; i++) {
-      fetchTask(tasks[i]);
-    }
-  });
-});
-
-var a = 'hehehe';
-function getTasks() {
-  let promise = new Promise( (resolve, reject) => {
-    Api.getTasks( (tasks) => {
-      resolve(JSON.parse(tasks));
-    });
-  });
-  return promise;
-}
-
-function getTemplate() {
-  let promise = new Promise( (resolve, reject) => {
-    let http = new XMLHttpRequest();
-
-    http.open('GET', '../handlebars/taskTpl.hbs', true);
-    http.send(null);
-
-    http.onreadystatechange = () => {
-      if (http.readyState === 4 && http.status === 200) {
-        resolve(http.responseText);
-      }
-    };
-
-    http.onerror = (e) => {
-      reject(e);
-    };
-  });
-
-  return promise;
-}
-
-function fetchTask(task) {
-  let typeSection = document.getElementsByClassName(task.type)[0],
-      errMsg = 'Required fields are missing.',
-      okMsg = 'Everything went fine!';
-
-  if (!task.name || !task.author || !task.type) {
-    return errMsg;
-  }
-  else {
-    if (!typeSection) return;
-
-    let taskHtml = taskTpl(task),
-        currentTask = document.createElement('div');
-    
-    currentTask.setAttribute('data-taskid', task._id);
-    currentTask.classList.add('task');
-    currentTask.innerHTML = taskHtml;
-
-    typeSection.appendChild(currentTask);
-    return okMsg;
-  }
-}
-
-function addTask(options) {
-  let promise = new Promise( (resolve, reject) => {
-    Api.addTask(options, (task) => {
-      fetchTask(JSON.parse(task));
-      return 'Task succesfully added.';
-    });
-  });
-  toggleDialog();
-}
-
-function editTask(taskId, taskObj) {
-  let promise = new Promise( (resolve, reject) => {
-    Api.editTask(taskId, taskObj, {}, (task) => {
-      updateTask(JSON.parse(task));
-
-      //end of editing 
-      isEditing = false;
-    });
-  });
-  toggleDialog();
-}
-
-function updateTask(task) {
-  let taskEl = document.querySelectorAll('[data-taskid="' + task._id + '"]')[0],
-      formerType = taskEl.closest('.board').classList[1];
-
-  // if changed task type - move task to different board
-  if (formerType !== task.type) {
-    taskEl.parentElement.removeChild(taskEl);  
-    fetchTask(task); 
-  }
-  else {
-    taskEl.getElementsByClassName('task-name')[0].textContent = task.name;
-    taskEl.getElementsByClassName('task-author')[0].textContent = task.author;
-    taskEl.getElementsByClassName('task-desc')[0].textContent = task.description;
-  }
-}
-
-window.toggleDialog = (taskObj) => {
- let dialog = document.getElementsByClassName('add-task-dialog')[0],
-     inputs = document.querySelector(".properties input"),
-     editTitle = 'Edit Task',
-     addTaskTitle = 'Add New Task';
-
- if (dialog.classList.contains('hidden')) {
-    dialog.classList.remove('hidden');
- }
- else {
-    dialog.classList.add('hidden');
-
-    for (let i = 0 ; i < inputs.length ; i++) {
-      inputs[i].value = '';
-    }
-  }
-
-  let titleEl = document.querySelector(".add-task-dialog h2");
-  //while editing
-  if (taskObj) {
-    let {taskName, taskAuthor, taskDesc, taskType} = taskObj;
-    document.querySelector(".name-input input").value = taskName;
-    document.querySelector(".author-input input").value = taskAuthor;
-    document.querySelector(".description-input input").value = taskDesc;
-    document.getElementById(taskType).checked = true;
-    titleEl.innerHTML = editTitle;
-  }
-  //while creating new task
-  else {
-    titleEl.innerHTML = addTaskTitle;
-  }
-};
-
-window.getNewTaskOptions = () => {
-  let taskName = ocument.querySelector(".name-input input").value,
-      taskAuthor = document.querySelector(".author-input input").value,
-      taskDesc = document.querySelector(".description-input input").value,
-      typeCheckbox = document.querySelector(".task-type-form input"),
-      taskType;
-
-  for (let i = 0 ; i < typeCheckbox.length ; i++) {
-    if (typeCheckbox[i].checked) {
-      taskType = typeCheckbox[i].value;
-    }
-  }
-
-  //all fields must be filled
-  // TODO - visible alert that you haven't filled all inputs
-  if (!taskName || !taskAuthor || !taskDesc) return;
-
-  let taskObject = {
-    name: taskName,
-    author: taskAuthor,
-    type: taskType,
-    description: taskDesc
-  };
-
-  if (isEditing) {
-    editTask(taskId, taskObject);
-  }
-  else {
-    addTask(taskObject);
-  }
-};
-
-window.deleteTask = (obj) => {
-  let taskEl = obj.parentElement,
-      taskId = taskEl.getAttribute('data-taskid');
-
-  let promise = new Promise( (resolve, reject) => {
-    Api.deleteTask(taskId, (task) => {
-      taskEl.parentElement.removeChild(taskEl);  
-      return 'Task succesfully deleted.';
-    });
-  });
-};
-
-window.showEditDialog = (obj) => {
-  let taskEl = obj.parentElement,
-      taskObj = {
-        taskName: taskEl.getElementsByClassName('task-name')[0].textContent,
-        taskAuthor: taskEl.getElementsByClassName('task-author')[0].textContent,
-        taskDesc: taskEl.getElementsByClassName('task-desc')[0].textContent,
-        taskType: taskEl.closest('.board').classList[1]
-      }
-  
-  taskId = taskEl.getAttribute('data-taskid');
-  isEditing = true;
-  toggleDialog(taskObj);
-};
-module.exports = showEditDialog;
-export {fetchTask, addTask, toggleDialog, editTask};
-
+window.toggleDialog = app.toggleDialog;
+window.getNewTaskOptions = app.getNewTaskOptions;
+window.showEditDialog = app.showEditDialog;
+window.deleteTask = app.deleteTask;
